requirement.txt
asgiref==3.7.2
Django==4.2
django-bootstrap5==23.4
django-filter==23.5
django-taggit==5.0.1
djangorestframework==3.14.0
Markdown==3.5.2
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
